tickbox/pre-commit: Skip some tests if not running SLOW

+++ tickbox/precommit/30-test.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 set -ueo pipefail
-cd "$TICKBOX_TEMPDIR/work"
-export CARGO_TARGET_DIR="$TICKBOX_CWD/target/${TICKBOX_BRANCH}.test.normal"
-cargo test --workspace
-if [[ ${CLEANUP:-} = true ]]; then
-        rm -fr "${CARGO_TARGET_DIR?}"
+if [[ ${SLOW:-} = "true" ]]; then
+        cd "$TICKBOX_TEMPDIR/work"
+        export CARGO_TARGET_DIR="$TICKBOX_CWD/target/${TICKBOX_BRANCH}.test.normal"
+        cargo test --workspace
+        if [[ ${CLEANUP:-} = true ]]; then
+                rm -fr "${CARGO_TARGET_DIR?}"
+        fi
 fi
